+++ package.json
@@ -14,6 +14,7 @@
     "@types/platform": "^1.3.2",
     "@types/react": "^16.9.43",
     "@types/react-dom": "^16.9.0",
+    "@types/react-router-dom": "^5.1.5",
     "classnames": "^2.2.6",
     "downshift": "^6.0.0",
     "finalhandler": "^1.1.2",
@@ -23,9 +24,11 @@
     "platform": "^1.3.5",
     "react": "^16.13.1",
     "react-dev-utils": "^10.2.1",
+    "react-device-detect": "^1.13.1",
     "react-dom": "^16.13.1",
     "react-dropzone": "^11.0.2",
     "react-reveal": "^1.2.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "styled-jsx": "^3.3.0",
     "typescript": "^3.9.6"
